requirements.txt
anthropic>=0.39.0
music21>=9.1.0
Pillow>=10.0.0
opencv-python-headless>=4.8.0
numpy>=1.24.0
imagehash>=4.3.0
lxml>=4.9.0
click>=8.1.0
rich>=13.0.0
